Makefile: [fix] allow minor releases for RPMs
Previously, incrementing the Release number in stb-tester.spec.in to
anything other than 1 would cause the srpm make target to break because
the name of the produced RPM would not match that which was hardcoded to
the "src_rpm" variable.

This commit fixes the issue by substituting the minor release into the
RPM spec file in the same way that the major version is. Users wishing to
create minor releases can now either hardcode the increment into the
Makefile by changing the assignment to the `RELEASE` variable or simply
assign it on the `make` command line, e.g. `make rpm RELEASE=3`.