Makefile: add `sudo` specifically for yum-builddep in `make rpm`
`yum-builddep` requires super user privileges, which meant that users
would have to run `make rpm` as root user in order for it to succeed. This
commit adds a `sudo` specifically for the `yum-builddep` call to avoid
running the whole make recipe (and any of its dependencies) as the root
user, which would, for example, create the "VERSION" file with root privs.