stbt._draw_text: Enlarge the black rectangle to create a 2px margin
It is easier to read the text if the black rectangle does not end at
the outlines the text. E.g. the letter 'E' on a white background is
still difficult to read because the top and bottom horizontal lines are
right at the the edges of the black rectangle.