stbt run: Treat AssertionErrors as test failures, not test errors
This brings stb-tester in line with traditional testing frameworks like
nose and pytest. Along with the introduction of `match()` in 33e8e356
and (in a future commit) `wait_until()`, this is part of a drive to
provide a more composible API. You can say things like:

* `assert match("xyz.png")`
* `assert not match("xyz.png")`
* `assert match("a.png") and match("b.png") and not match("c.png")`
* `assert wait_until(lambda: not match("xyz.png"))`
* `assert match_text("Main menu")`
* etc.

There are drawbacks to using `assert` instead of `wait_for_match`:

* The exception message won't contain the reason why the match
  failed (unless you specify it as a second parameter to `assert`,
  which is tedious and we don't expect you to do it), and
* The exception won't have the offending video-frame attached (so the
  screenshot that `stbt batch run` saves alongside the failing test
  logs will be a few frames later than the frame that actually caused
  the test to fail).

We hope to solve both problems at some point in the future: the first
one by using pytest and its assertion introspection[1] (which rewrites
python modules on import to provide the additional information on
assertion failures), and the second one by extending pytest's
module-rewriting to attach the video frame to the AssertionError.

[1] http://pytest.org/latest/assert.html#advanced-assertion-introspection