Add stbt.wait_until helper function
`wait_until` runs a function or lambda until it succeeds, or until a
timeout. See the function documentation in README.rst and
release-notes.md for details.

We have been using this function in one of our clients' test-pack
codebase for 6+ months now, and we use it in almost every single
testcase.