make doc: Fix formatting in manpage
The unquoted "callable_" was causing the following error:

    <stdin>:900: (ERROR/3) Unknown target name: "callable".

(reStructuredText uses the trailing "_" to indicate intra-document
links.)

This isn't in the docstring (where I could fix it by adding the quotes
in the python source file). It is in the function signature, which is
generated by our convoluted "api-doc.sh" script. If we were using Sphinx
it would have taken care of this for us.