stbt batch summary report: Hardcode column types for tablequeryjs
Fixes #257.

tablequeryjs (the library we use for browser-side filtering of testrun
results) looks at the table's first row (after the headings) to
determine what each column's data type is. For this it uses momentjs,
a third-party library bundled by tablequeryjs.

Sometimes it gets it wrong. For example it thinks that a particular git
sha is a timestamp, and so you can't search for textual matches on the
"git sha" column.

Another (more serious) bug is that if the content of a table-cell
(in the first row after the headings) is longer than ~150 characters,
there's a null-dereference error somewhere deep in momentjs and the
entire page breaks, not just the search.

This seems to fix it. I'm not interested in writing an automated
regression test for it. At some point hopefully soon we're going to
implement server-side filtering (for other reasons, not specifically to
get rid of tablequeryjs, but it's likely that we won't be using
tablequeryjs at all so there's no point in spending too much effort on
this).

The hidden row has to be in a separate <thead> because sorttable (the
library that allows client-side sorting by clicking on a column header)
doesn't like a <thead> with more than one row -- it explicitly has this
line of code:[1]

    if (table.tHead.rows.length != 1) return; // can't cope with two header rows

The hidden row can't be in the <tbody> because then tablequery will
un-hide it when you clear a search.

[1] http://www.kryogenix.org/code/browser/sorttable/sorttable.js