Replace references to github.com/{drothlis => stb-tester}/stb-tester
We have moved the official stb-tester github repository from
  https://github.com/drothlis/stb-tester
to
  https://github.com/stb-tester/stb-tester

This was done to give greater visibility to the other repositories that
we will also be adding to https://github.com/stb-tester at some point in
the future (such as stb-tester-service[1]) and to make it clearer which
is the canonical version out of all the github forks.

Github will automatically redirect from the old location to the new
location, but for clarity it's worth updating all mentions of the
location anyway.

[1] http://stb-tester.com/blog/2014/08/23/upcoming-features-stateless-service.html