stbt power: Implement in Python rather than bash
We've been meaning to [move to Python] for the power controllers for a
while as shell isn't really that good a fit for argument parsing, URI
matching and making HTTP requests.  Also it will make it easier to access
this functionality from the Python scripts themselves.

With this commit all of the real functionality is still in `stbt-power.sh`
but the actual `stbt power` command itself is in Python which opens up a
migration path to allow the existing supported power controllers to be
ported to Python.

Note: none of the currently supported power controllers have any tests at
all and the command itself is barely documented, so it's difficult to be
sure that we haven't broken anything.

This is in preparation for adding support for adding support for
[Aviosys USB NET POWER 8800 Pro] power controllers.

[move to Python]: https://groups.google.com/d/msg/stb-tester/0wFNSIimHn0/317YMmc4CesJ
[Aviosys USB NET POWER 8800 Pro]: http://www.aviosys.com/8800.html