stbt python functions: Raise `ValueError` instead of `AssertionError`
Since d91cad41, `AssertionError`s are treated as test failures, not
errors; so we can't use `assert` to validate user input.

There are still a couple of `assert`s in `stbt.Display`'s GStreamer
error-handling callbacks, but those are genuine this-should-never-happen
assertions so I think that's ok.

This doesn't address the problem of third-party libraries (called by the
user's test script) using `assert` in a way that doesn't align with
`stbt run`'s interpretation.