stbt run: Print details of failed assert statement
`assert` takes an optional error message, and if you don't specify it,
then the `str(AssertionError)` is blank. In the `stbt batch` report,
you just see "AssertionError:" as the failure reason.

With this commit, the failure reason will show the line of code that
raised the `AssertionError`, but only if the `assert` statement was not
given an explicit error message.

This isn't perfect: For multi-line `assert` statements, you'll only see
the last line. But it's still very useful in most cases.