Drop "0." from version numbering scheme for stb-tester 22 onwards
We've changed the release numbering here, dropping the "0." from the
beginning of the version number.  stb-tester has been usable software that
we've cared about backwards compatibility for years now so "0." simply
doesn't make sense.

Technically it is correct to bump the major revision on every release in
accordance with "semantic versioning" as we are not completely strict about
backwards-compatibility between releases.