stbt.{ocr,match_text}: Be more explicit about the default value of `lang`
Strings are immutable, so there's no need to use this workaround to
Python's "mutable default arguments" gotcha.

This helps to make the function signature more explicit in the generated
API documentation.

Note that the `_tesseract` internal helper function doesn't need to
specify the defaults again, because all callers pass in a value
explicitly. So I've removed those defaults to prevent them getting out
of sync in the future. Technically this creates a backwards
incompatibility if anyone called `ocr(lang=None)` (that is, explicitly
setting `lang` to `None`) but I doubt it'll actually affect anyone.