stbt.OcrMode: Is now an `enum.IntEnum`
This is so that the generated API documentation for `ocr`'s signature
says "mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD" instead of "mode=3".

I override `OcrMode.__repr__` so that the documentation says
"OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD" instead of
"<OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD: 3>". It seems to me that this
new `__repr__` is actually more in the spirit of repr -- that is, you
could paste the string it outputs into a python interpreter and get
back the corresponding object.

`enum` is built into python 3.4; the `python-enum` package backports it
to python 2.