release notes: Fix hyperlinks to specific versions
When we used to host the release notes on stb-tester.com, we had a
custom markdown-to-html script that inserted anchors named "0.21" where
the 0.21 heading started.

Now the release notes are just hosted on github's source code viewer,
which automatically displays markdown as html and creates anchors
matching the text of the heading. I've made the heading be just
"0.21" (instead of "0.21: Composable API, etc") so that old hyperlinks
will once again point to the right place within the document.