Fix ubuntu packaging, broken in 531be649
After 531be649, `extra/debian/changelog` said "stb-tester (22-1)"
instead of "stb-tester (22-@RELEASE@)" (@RELEASE@ was supposed to
be filled in later by `extra/debian/build-source-package.sh`).

This change moves the creation of `changelog` from `changelog.in`
entirely out of the Makefile and into `build-source-package.sh`. (Before
it was partly in the Makefile and partly in `build-source-package.sh`.)