extra/fedora/test-rpm.sh: Allow also installing stb-tester-camera rpm
Run it like this:

    extra/fedora/test-rpm.sh stb-tester-{,camera-}-$version-1.fc20.x86_64.rpm

Currently this isn't terribly useful because to install
stb-tester-camera you need gstreamer1-libav, which isn't available in
the Fedora repositories, and this script doesn't install it for you. But
I'm committing it anyway so I don't do the work again at the next
stb-tester release only to find out that I had forgotten about the libav
thing.