extra/fedora/copr-publish.sh: Only run `copr-cli` in Fedora container
The `copr-cli` tool is only available in Fedora, so it needs to be run
in a Fedora container. Now that the `git clone` is done on the host, I
can use git's ssh transport (I don't have to type in a password) without
having to figure out how to get the credentials into the container.