test_clock_visualisation: Fix skipping logic broken in 68983e0
In 68983e0 we stopped running integration tests with `PYTHONPATH` set.

test_clock_visualisation requires tesseract 3.03, and at the beginning of
the test we check this by running a little Python snippet.  The return code
of this snippet was to distinguish between tesseract versions, but it
didn't make a distinction between version 3.02 being installed and any
other miscellaneous error.  This silently caused this test to be skipped.

This commit fixes this issue.  Thankfully the test still passes.