Tests: Fix race in test_that_stbt_batch_reports_results_directory
In `test_that_stbt_batch_reports_results_directory` we use socat to fake
out the tracing socket.  By default `socat` has a receive timeout of 0.5s
after which it will assume a socket is closed.  From the socat man page:

>     -t<timeout>
>         When  one  channel has reached EOF, the write part of the other
>         channel is shut down. Then,  socat  waits  <timeout>  [timeval]
>         seconds  before terminating. Default is 0.5 seconds. This time‐
>         out only applies to addresses where write and read part can  be
>         closed independently. When during the timeout interval the read
>         part gives EOF, socat terminates without awaiting the timeout.

This didn't trip the test up when running it locally, but Travis is slower
so we would see the failures there.

This commit increases the timeout to 10s, fixing the issue.  We also
increase the debugging level of socat to make it easier to understand
issues like this in the future.