draw_text: Fix race conditions
Previously `push_sample` was rather confusing.  We were using the local
variable `texts` as if it we had sole ownership of it when actually it was
an alias for `self.text_annotations` and was mutable shared state modified
in other threads by `draw_text`.

This commit fixes these issues and improves symmetry between the way that
matches and text are dealt with.