stbt-run: Handle failures registered by uitests/common.SurviveFailure
We would only like to see ignored failures if the test didn't raise
any uncaught exception, i.e. it did not finish with a higer priority
failure. As stb-tester/stbt-run sets the script's exit code, it is not
possible to get the exit status in an uitests module without serious
hacking.

Move the code reporting ignored failures to stbt-run in the YouView
stb-tester fork.

NOTE: This implementation has a regression compared to the previous
solution in uitests using `atexit`: it doesn't report ignored failures
if the test script exits with `sys.exit`. As we don't have any test that
calls `sys.exit`, and we shouldn't have any in the future because we
should always raise an exception to indicate a test error or failure,
it shouldn't cause problems.