Makefile: Facilitate versioning internal stb-tester releases
stb-tester's way of generating version number makes it difficult to
publish minor internal releases frequently, because it requires us to
create a new annotated tag for every release. Hard-code VERSION and
RELEASE and so they can be manually incremented to create new releases,
just like we do in the stb-tester-slave.spec file. Also include the
release number in `stbt --version` so that we can easily keep track of
minor releases.