Makefile: Don't build/install man page if 'rst2man' isn't installed
We have some users who want to install the stb-tester python libraries
on OS X just so that they can write scripts that use `stbt.Region` and
other utilities that don't require GStreamer, or so that they can run
pylint over their test scripts. They don't need to *run* their scripts
on OS X, because the use the stb-tester ONE appliance[1] for that.

This change makes it easier for those users to install the stbt python
libraries, because OS X doesn't have `rst2man`; furthermore `pip install
docutils` installs it as `rst2man.py` and puts it somewhere that isn't
on the default PATH.

[1] http://stb-tester.com/stb-tester-one