stbt,-run,-batch: allow users to register non-fatal failures mid-test
It is sometimes the case that a test script encounters an error during
execution which the user can anticipate and work around, but would also
like to have a record of encountering. With this commit, the user can
register such a failure with `stbt` which `stbt` then reports during the
exit of `stbt-run` and can also use to fail an otherwise passing test.

The added selftests give comprehensive if trivial ideas on the use cases.
More concrete examples can be seen in the module docstring for
`_stbt/failure_reporting.py`.

For the reporting-after-test-completion to work as expected, `stbt-run`
required separate handling for the registered failures for both cases when
the executing test did and didn't propagate an unhandled exception up to
stbt-run (i.e. it failed in a way the user wasn't expecting). In cases
where there was an unhandled exception, `stbt-run` with report any failures
registered but these failures don't affect *how* `stbt-run` fails itself.
Where the executing test *doesn't* fail, any registered failures will
affect how `stbt-run` exists.

Given that `stbt-run` already maps `stbt.UITestFailures` to exit code 1
and any other exceptions to exit code 2, with the intention being that
"UITestFailures are more important/interesting", it was decided that where
multiple failures had been registered then the lowest exit code from those
failures would take priority. It was also decided that when a single
failure with an exception was registered then that exception would be
raise rather than a proxy-exception, as this makes the failure much faster
to analyse because the information is immediately present. However,
printing the details of multiple exceptions would clutter the summary
output.