librification: Introduce `DeviceUnderTest` class
This is an object that subsumes the previous global state variables
`_display` and `_control`.  Now all functions that require access to the
display or the control have become methods of `DeviceUnderTest`, and
access them via `self`.  `stbt/__init__.py` just forwards calls onto it's
global `DeviceUnderTest` instance named `_dut`.

This is a part of "librificaiton" work, better allowing stbt to be used as
a library, by removing side-effects.