librification: Make `DeviceUnderTest` a `contextmanager`
This makes separates the lifetime of the `DeviceUnderTest` from the time
it's actually active.  This will be useful to re-use a `DeviceUnderTest`
object multiple times and tidies up the `init_run`/`teardown` dance.

This is a part of "librificaiton" work, better allowing stbt to be used as
a library, by removing side-effects.