Rewrite `stbt batch run` in Python
This is in preparation for adding additional flags to `stbt batch run`
controlling the order in which tests are run.  The implementation in
Python should be easier to extend.

Note: This implementation still has a regression (in the Ctrl-C handling
vs the original implementation in bash.  This will be fixed in
subsequent commits.