stbt batch run: Run child process in a new process group
After the first `Ctrl-C` `stbt batch run` will not start any additional
tests, but will not interrupt the currently running test.  You need to
send a second `Ctrl-C` to interrupt the current test.

This behaviour was removed during my previous refactoring of the
`stbt batch run` code (while it was written in bash).  This commit
reintroduces that behaviour now that it is written is Python and is easier
to implement.

When you press `Ctrl-C` in a terminal the tty will send SIGINT to all the
processes in the foreground process group.  With this change we run the
child `stbt batch run` process in a separate process group to the
`stbt batch` process.  This gives the parent more fine-grained control over
which signals to forward to its children when, making the "double `Ctrl-C`"
behaviour easier to implement and making it easier to kill all the child
processes without complexity like `killtree`.