Split stbt-batch run up into scheduler and runner
* The scheduler (`stbt-batch.d/run` decides which tests to run and
  invokes the runner.
* The (private) runner (`stbt-batch.d/run-one`) runs the tests and all
  the associated pre_run, post_run, classify, recover, etc. scripts.

I've split these into seperate executables and processes to make the
responsibilities and lines of communication between them clearer.  This
will make it easier to port one or the other piecemeal to Python.

Note: This commit removes `stbt batch run`'s feature that on Ctrl-C
the actual test is only interrputed after the second Ctrl-C.  It proved
to be too difficult to implement in bash under this new model, but the
feature will be reimplemented in two commits time, once the scheduler
is reimplemented in Python.