stbt lint: Check that `wait_until` argument is callable
The intention is to catch common mistakes like
`wait_until(is_screen_black())` when you meant
`wait_until(is_screen_black)`.

Calling a function *can* be valid, if that function returns another
function. Unfortunately pylint's inference doesn't always succeed at
figuring out what the function is. If we can't figure it out, we don't
complain even if it looks like a function call. As the unit test shows,
we don't actually complain about the above example. It will still catch
some cases though, so it's still useful.

An alternative approach would be to complain if it's a function call and
we fail to infer the return value.