stbt lint: Stricter check that `wait_until` argument is callable
Passing a function-call as the argument to `wait_until` *can* be valid
--if that function returns another function-- but it's fairly rare to
actually do so. If pylint fails to infer the return type from that
function, let's complain about it. This runs the risk of introducing
false positives, but that's better than false negatives, and you can
always silence false positives with `# pylint:disable=E7003`.