New API stbt.Region.replace
This is similar to `Region.extend`, but it takes absolute coordinates
within the image instead of adjusting by a relative number of pixels.

It just delegates to the private `namedtuple._replace` method (`Region`
is a subclass of `namedtuple`). We have found it quite useful in some
of the test cases we write, and this saves us from having to use a
private method.