Travis 14.04 support: Set $PATH to standard for tests
Travis comes with it's own python interpreter installed into /opt.  This
is separate from the system installation of python so we can't use system
installed python libraries with that version of python. This is no good for
us as we want to use the system libraries opencv and gstreamer which we've
`apt-get install`ed.

This issue was causing our tests to fail with:

> ImportError: No module named cv2

This commit removes the custom python from PATH so we use the system one
instead.