stbt-run: Save last used frame on failure as screenshot.png
stbt-run knows that a test has failed/errored if the test script raises
an exception.  When `wait_for_match`, `wait_for_motion`, etc. fail
they attach the frame that caused the failure to the exception that they
raise.  stbt-run is expecting this so can save the screenshot for later
inspection and debugging.

However: if users are using the more modern style of:

    assert wait_until(lambda: match(...))

then the exception is raised by the assert, which has no knowledge of
the frame that caused the assert.  In this case `stbt batch run` will
grab a screenshot of whatever happens to be on screen when the test
finishes.  This screenshot often doesn't exhibit the issue that caused
the test to fail/error in the first place.

This commit makes the assumption that the last frame that the test script
actually used is the one that exhibited the issue and saves that instead.