stbt.init_run(): Don't leave _dut in inconsistent state on error
This is unlikely to have any real world effect but makes `init_run` better
behaved.  Previously if `DeviceUnderTest.__enter__()` raised an exception
`stbt._dut` would be in an error state and `init_run` would propogate that
exception through.  Now `init_run` either it throws an exception and has
no effect or it succeeds and `_dut` is set appropriately.