Work around regression in PyGObject 3.13
Preventing stb-tester from working on Ubuntu >= 14.10 and Fedora >= 21.

`map_gst_sample` was failing with:

    Traceback(most recent call last):
      File"/usr/libexec/stbt/stbt/__init__.py", line 1920, in on_new_sample
        self.push_sample(sample)
      File"/usr/libexec/stbt/stbt/__init__.py", line 1984, in push_sample
        with _numpy_from_sample(sample)as img:
      File"/usr/lib64/python2.7/contextlib.py", line 17, in __enter__
        returnself.gen.next()
      File"/usr/libexec/stbt/stbt/__init__.py", line 1708, in _numpy_from_sample
        with map_gst_buffer(sample.get_buffer(), flags) as buf:
      File"/usr/lib64/python2.7/contextlib.py", line 17, in __enter__
        return self.gen.next()
      File"/usr/libexec/stbt/_stbt/gst_hacks.py", line 47, in map_gst_buffer
        "Writable array requested but buffer is not writeable")
    ValueError: Writable array requested but buffer is not writeable

PyGObject changed the behaviour of `Gst.Sample.get_buffer` to take a
reference to the buffer it returns with `gst_mini_object_ref`.  to
prevent use-after-free bugs if the `Gst.Buffer` is used after the
`Gst.Sample` goes out of scope.  Unfortunately `gst_mini_object_ref` isn't
just a reference-counting increment, it also marks the object as read-only.
This makes it impossible to get a writable handle to the `Gst.Buffer` owned
by the `Gst.Sample`.

In this commit we work around this by using the C API directly via ctypes
as we had to for `map_gst_buffer()` (see f0c5b31).

See the upstream bug ticket:
https://bugzilla.gnome.org/show_bug.cgi?id=736896

Fixes: #305