Display.push_sample: write template name above MatchResult annotation
Debugging/Analysing a test failure from the video alone can be difficult
because the template matching regions are not annotated, and cross
referencing with the stderr log is all but necessary in order to identify
what was being searched for at the time.

This commit reduces the cross-referencing required by labelling the
match annotations with the template name being searched for. This helps
to quickly identify what activity is going on at any given time.

Note that, as with the match annotations themselves, this will not affect
`stbt.match` calls which pass in their own frame, unless that frame is
provided by iterating over `stbt._display.gst_samples()` (note also that
`stbt._display.gst_samples().next()` *does not* work!)