core: Refactor MainLoop handling
...to allow multiple `Display` objects to co-exist.

Without this change tearing down a Display object would exit the main-loop
that a second `Display` was using.  As a result the EOS message would not
be received from the sink-pipeline of the second `Display` during teardown
causing a delay and potentially incomplete teardown.

With this change we have a main loop per `Display` object so different
`Display` objects aren't sharing the same state.

In the future it might be a good idea to share a main loop between
different `Display` objects, or with further librification, other parts of
the application.