core: Display teardown: Put sink-pipeline in NULL state before destruction
This is useful if EOS failed for any reason.  I was having some EOS
troubles where X windows weren't disappearing and this helps.

This is the right thing to do anyway and I'm surprised that GStreamer
hasn't been complaining about us destroying a pipeline that is not in the
NULL state.