core: Avoid ValueError on teardown
If we recieve a GStreamer error message during teardown it may be serviced
after `Display.pipeline` has been set to `None` resulting in an
`Exception` that we no-longer care about and is only caused by debug code.

This is probably only possible now that we're not calling
`mainloop.exit()` from the error handler.