.travis.yml: Install python-mock & pysnmp from apt, not pip
Now that we're using Ubuntu 14.04 on travis there's no reason not to.

The versions in the 14.04 repos match what I currently have installed
locally via pip.

pysnmp was available at a slightly older version in 12.04 as well -- I'm
not sure if it needed to be installed via pip in the first place.
python-mock was a much older version on 12.04.

Regarding the remaining packages that we still install with pip:

* python-rednose is only available via apt from 14.10 onwards.
* isort is only available from 15.04 onwards.
* pylint and astroid (a library used by pylint) are at an older version
  on 14.04. They break compatibility so much with each release that it's
  worth pinning the version, anyway.