core: Tear down the sink pipeline even if `NoVideo`
This if statement is left-over from aee8f04 (2013-07-02) when we didn't
have seperate source and sink pipelines.  It was introduced to avoid a
hang on teardown due to teardown during a pipeline restart.

Now that there are seperate source and sink pipelines it doesn't make
sense to conditionally teardown the sink pipeline based on whether the
source pipeline is having problems.