Makefile: Simplify make install rule
as a whole we follow a convention where the location in the source
directory matches the location in the installed directory.  In this
commit we simplify the make install rule taking advantage of this
convention.

This is in preparation for adding `stbt virtual-stb` which will
introduce yet another directory and yet more files to install.