stbt-run: Fix printing unicode tracebacks
Previously raising a exception with a unicode message would be backslash
escaped when printed to stderr:

    $ cat example.py
    # coding: utf-8
    assert False, u"Röthlis"

    $ stbt run example.py
    FAIL: example.py: AssertionError: Röthlis
    Traceback (most recent call last):
      File "./stbt-run", line 107, in <module>
        execfile(_filename)
      File "/dev/stdin", line 2, in <module>
        assert False, u"Röthlis"
    AssertionError: R\xf6thlis

This is not very helpful for readability, nor indexing and searching,
particularly if the messages are in Chinese.  Now we print it as utf-8:

    $ stbt run example.py
    FAIL: example.py: AssertionError: Röthlis
    Traceback (most recent call last):
      File "./stbt-run", line 107, in <module>
        execfile(_filename)
      File "/dev/stdin", line 2, in <module>
        assert False, u"Röthlis"
    AssertionError: Röthlis

In the tests I've had to grep out the libdc1394 errors as they occur on
Ubuntu 14.04 (and thus on Travis) but not on later distros.