Add extra/debian/test-deb.sh instead of `make deb` recipe
For consistency with `extra/fedora/test-rpm.sh` and to keep our Makefile
simpler.

This is based on `extra/fedora/test-rpm.sh` so it does more than the
`make deb` recipe did: It allows you to specify any deb to test, and it
displays the output of `stbt --version`, `stbt --help`, and `man stbt`
in addition to running the self-tests.