librification: Remove global state from _stbt
The `get/set_config` mechanism is a way of storing state in configuration
files.  This state is global, meaning that you couldn't have different
devices under test configured separately in the same process.

This commit fixes that issue by pulling a config object into the
`DeviceUnderTest` class.  This is a fairly blunt approach but it works for
now.  Perhaps in the future the specific keys that the `DeviceUnderTest`
class needs can be directly injected.