stbt-run/state_watch: don't monopolise `sys.settrace`
Previously, whether `--save-trace` was given or not, a tracer was created
and started and `sys.settrace` was called. When porting other changes
forward on a more recent version of `stbt`, including our own custom
tracing option, I found that the existing tracer was interfering with it
because `sys.settrace` was always called immediately before a test was
executed, overriding our own tracer.

This commit modifies `stbt-run` and `state_watch` so that the option
really is optional, and if not given won't cause side effects. This
includes refactoring the trace setup code out of `stbt-run` and into
`state_watch`, which helps to tidy up the code. The previous functionality
is otherwise maintained, other than that to save the trace to the default
socket you *must* set `--save-trace` (without any value) (and then follow
it with `--` to prevent the test-script arg being interpreted as the
"FILE" value for `--save-trace` if no other options are given).