man page: Remove obsolete/duplicate information
A lot of this stuff isn't reference material, and is covered in a more
up-to-date fashion in the wiki, particularly in the following articles:

* https://github.com/stb-tester/stb-tester/wiki/Introducing-stb-tester
* https://github.com/stb-tester/stb-tester/wiki/Getting-started-with-stb-tester
* https://github.com/stb-tester/stb-tester/wiki/Video-capture-hardware-for-DIY-stb-tester-rigs
* https://github.com/stb-tester/stb-tester/wiki/Infrared-hardware-for-DIY-stb-tester-rigs

HARDWARE REQUIREMENTS: "The Hauppauge HD PVR works well" -- what!? This
must've been terribly confusing for newcomers. Link to the wiki instead,
and suggest the stb-tester ONE.

TEST SCRIPT BEST PRACTICES: Doesn't really belong here, and it wasn't a
particularly great list in the first place -- no great loss.

AUTHORS: man-pages(7) says "Use of an AUTHORS section is strongly
discouraged".